requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
langchain
faiss-cpu
chromadb
langchain_google_genai